refactor: 검증 관련 폴더 생성
